pow-migration: Move some functionality to the library
Move some functionality from the `pow-migration` binary into its
library such that some of these components can be easily tested from
integration tests.
